AS7-6325 pure refactoring
