+++ connector/src/test/java/org/jboss/as/connector/subsystems/complextestcases/ComplexJcaSubsystemTestCase.java
@@ -19,12 +19,10 @@
  * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
-package org.jboss.as.connector.subsystems.jca;
+package org.jboss.as.connector.subsystems.complextestcases;
 
 
-import static org.junit.Assert.*;
-
-import org.junit.Ignore;
+import org.jboss.as.connector.subsystems.jca.JcaExtension;
 import org.junit.Test;
 
 /**
