+++ connector/src/test/java/org/jboss/as/connector/subsystems/complextestcases/ParseUtils.java
@@ -19,7 +19,7 @@
 * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
 * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
 */
-package org.jboss.as.connector.subsystems.jca;
+package org.jboss.as.connector.subsystems.complextestcases;
 
 import java.util.Enumeration;
 import java.util.Properties;
