+++ connector/src/test/java/org/jboss/as/connector/subsystems/jca/JcaSubsystemTestCase.java
@@ -39,25 +39,7 @@ public class JcaSubsystemTestCase extends AbstractSubsystemBaseTest {
     @Override
     protected String getSubsystemXml() throws IOException {
         //TODO: This is copied from standalone.xml you may want to try more combinations
-        return
-            "<subsystem xmlns=\"urn:jboss:domain:jca:1.0\">" +
-            "    <archive-validation enabled=\"false\" />" +
-            "    <bean-validation enabled=\"false\" />" +
-            "    <default-workmanager>" +
-            "        <short-running-threads>" +
-            "                <core-threads count=\"10\"/>" +
-            "                <queue-length count=\"30\"/>" +
-            "                <max-threads count=\"30\"/>" +
-            "                <keepalive-time time=\"10\" unit=\"seconds\"/>" +
-            "        </short-running-threads>" +
-            "        <long-running-threads>" +
-            "                <core-threads count=\"15\"/>" +
-            "                <queue-length count=\"10\"/>" +
-            "                <max-threads count=\"25\"/>" +
-            "                <keepalive-time time=\"10\" unit=\"seconds\"/>" +
-            "        </long-running-threads>" +
-            "    </default-workmanager>" +
-            "</subsystem>";
+        return readResource("jca.xml");
     }
 
     @Override
