+++ connector/src/test/resources/org/jboss/as/connector/subsystems/complextestcases/datasource.xml
