+++ connector/src/test/resources/org/jboss/as/connector/subsystems/complextestcases/ra2.xml
