+++ connector/src/test/resources/org/jboss/as/connector/subsystems/datasources/datasources-full.xml
