+++ connector/src/test/resources/org/jboss/as/connector/subsystems/datasources/datasources-minimal.xml
