+++ connector/src/test/resources/org/jboss/as/connector/subsystems/resourceadapters/empty-resourceadapters.xml
@@ -1,0 +1,1 @@
+<subsystem xmlns="urn:jboss:domain:resource-adapters:1.0" />
